instructions in readme, error handling, concrete cache miss exception, code quality

+++ src/main/java/com/ncr/geocode/services/AddressService.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface AddressService {
 
     List<Address> getCachedAddresses();
-    Address get(float lat, float lon);
+    Address get(double lat, double lon);
 }
